phase3/backend/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
asyncpg
aiosqlite
pydantic
pydantic-settings
passlib
bcrypt
python-jose
cryptography
python-multipart
python-dotenv
httpx
requests
psycopg2-binary
alembic
pytest
pytest-asyncio
python-slugify
phonenumbers
email-validator
mcp
openai